feat: Add full Z-Machine v4+ object support
- Add version-aware property handling for v4+ games
  - Support 63 property defaults (v3 has 31)
  - Support 6-bit property numbers (v3 has 5-bit)
  - Support 2-byte property size format
  - Support properties up to 64 bytes (v3 max 8)

- Update object table handling for v4+ games
  - Support 14-byte object entries (v3 has 9)
  - Support 2-byte object numbers (v3 has 1-byte)
  - Support 48 attributes (v3 has 32)
  - Update parent/sibling/child to use u16

- Add helper function get_property_info() for version-aware property parsing
- Update all object-related functions to handle both v3 and v4+ formats
- Fix "Property 50 not found for object 17" error in AMFV

This enables support for v4+ games like A Mind Forever Voyaging.

+++ src/interpreter.rs
@@ -873,7 +873,7 @@ impl Interpreter {
                 // get_sibling
                 let sibling = self.vm.get_sibling(operand)?;
                 if let Some(store_var) = inst.store_var {
-                    self.vm.write_variable(store_var, sibling as u16)?;
+                    self.vm.write_variable(store_var, sibling)?;
                 }
                 self.do_branch(inst, sibling != 0)
             }
@@ -881,7 +881,7 @@ impl Interpreter {
                 // get_child
                 let child = self.vm.get_child(operand)?;
                 if let Some(store_var) = inst.store_var {
-                    self.vm.write_variable(store_var, child as u16)?;
+                    self.vm.write_variable(store_var, child)?;
                 }
                 self.do_branch(inst, child != 0)
             }
@@ -894,7 +894,7 @@ impl Interpreter {
                 );
                 let parent = self.vm.get_parent(operand)?;
                 if let Some(store_var) = inst.store_var {
-                    self.vm.write_variable(store_var, parent as u16)?;
+                    self.vm.write_variable(store_var, parent)?;
                 }
                 Ok(ExecutionResult::Continue)
             }
@@ -1111,7 +1111,7 @@ impl Interpreter {
                 // jin
                 // Check if obj1 is inside obj2 (obj1's parent is obj2)
                 let parent = self.vm.get_parent(op1)?;
-                let condition = parent == op2 as u8;
+                let condition = parent == op2;
                 self.do_branch(inst, condition)
             }
             0x07 => {
